requirements.txt
fastapi 
uvicorn 
httpx 
sqlalchemy 
Pillow 
python-multipart 
psycopg2-binary